@@ -14,6 +14,9 @@ namespace VirtualCamera.Src
     {
         public override Edge[] Edges { get; set; }
         public override Polygon[] Polygons { get; set; }
+        private readonly Texture2D crazy_cat;
+        private readonly Texture2D colour;
+        public List<Vector3> VerticeNormals { get; set; }
 
         public Cuboid(float x, float y, float z, float width, float length, float height)
         {
@@ -31,19 +34,46 @@ namespace VirtualCamera.Src
 
             PerspectiveVertices = new Vector4[Vertices.Count];
             Pixels = new Vector3[Vertices.Count];
+
+            crazy_cat = GraphicsManager.content.Load<Texture2D>("cat3_256");
+            colour = GraphicsManager.content.Load<Texture2D>("yellow_256");
         }
 
         public override void GenerateEdgesAndPolygons()
         {
             Polygons = new Polygon[]
             {
-                new Polygon(Pixels[0], Pixels[1], Pixels[3], Pixels[2], Color.Orange), //0 - dol
-                new Polygon(Pixels[0], Pixels[1], Pixels[5], Pixels[4], Color.Green), //1 - front
-                new Polygon(Pixels[0], Pixels[2], Pixels[6], Pixels[4], Color.Red), //2 - lewo
-                new Polygon(Pixels[1], Pixels[3], Pixels[7], Pixels[5], Color.Yellow), //3 - prawo
-                new Polygon(Pixels[2], Pixels[3], Pixels[7], Pixels[6], Color.Blue), //4 - tyl
-                new Polygon(Pixels[4], Pixels[5], Pixels[7], Pixels[6], Color.Violet) //5 - gora
+                new Polygon(Pixels[0], Pixels[1], Pixels[3], Pixels[2], Color.Orange, crazy_cat), //0 - dol
+                new Polygon(Pixels[0], Pixels[1], Pixels[5], Pixels[4], Color.Green, colour), //1 - front
+                new Polygon(Pixels[0], Pixels[2], Pixels[6], Pixels[4], Color.Red, colour), //2 - lewo
+                new Polygon(Pixels[1], Pixels[3], Pixels[7], Pixels[5], Color.Yellow, colour), //3 - prawo
+                new Polygon(Pixels[2], Pixels[3], Pixels[7], Pixels[6], Color.Blue, colour), //4 - tyl
+                new Polygon(Pixels[4], Pixels[5], Pixels[7], Pixels[6], Color.Violet, colour) //5 - gora
             };
+
+            Polygons[0].SetPerspectiveVertices(PerspectiveVertices[0], PerspectiveVertices[1], PerspectiveVertices[3], PerspectiveVertices[2]);
+            Polygons[1].SetPerspectiveVertices(PerspectiveVertices[0], PerspectiveVertices[1], PerspectiveVertices[5], PerspectiveVertices[4]);
+            Polygons[2].SetPerspectiveVertices(PerspectiveVertices[0], PerspectiveVertices[2], PerspectiveVertices[6], PerspectiveVertices[4]);
+            Polygons[3].SetPerspectiveVertices(PerspectiveVertices[1], PerspectiveVertices[3], PerspectiveVertices[7], PerspectiveVertices[5]);
+            Polygons[4].SetPerspectiveVertices(PerspectiveVertices[2], PerspectiveVertices[3], PerspectiveVertices[7], PerspectiveVertices[6]);
+            Polygons[5].SetPerspectiveVertices(PerspectiveVertices[4], PerspectiveVertices[5], PerspectiveVertices[7], PerspectiveVertices[6]);
+
+            VerticeNormals = new();
+            VerticeNormals.Add(MeanNormal(Polygons[0].Normal, Polygons[1].Normal, Polygons[2].Normal));
+            VerticeNormals.Add(MeanNormal(Polygons[0].Normal, Polygons[1].Normal, Polygons[3].Normal));
+            VerticeNormals.Add(MeanNormal(Polygons[0].Normal, Polygons[2].Normal, Polygons[4].Normal));
+            VerticeNormals.Add(MeanNormal(Polygons[0].Normal, Polygons[3].Normal, Polygons[4].Normal));
+            VerticeNormals.Add(MeanNormal(Polygons[1].Normal, Polygons[2].Normal, Polygons[5].Normal));
+            VerticeNormals.Add(MeanNormal(Polygons[1].Normal, Polygons[3].Normal, Polygons[5].Normal));
+            VerticeNormals.Add(MeanNormal(Polygons[2].Normal, Polygons[4].Normal, Polygons[5].Normal));
+            VerticeNormals.Add(MeanNormal(Polygons[3].Normal, Polygons[4].Normal, Polygons[5].Normal));
+
+            Polygons[0].SetLights(VerticeNormals[0], VerticeNormals[1], VerticeNormals[3], VerticeNormals[2]);
+            Polygons[1].SetLights(VerticeNormals[0], VerticeNormals[1], VerticeNormals[5], VerticeNormals[4]);
+            Polygons[2].SetLights(VerticeNormals[0], VerticeNormals[2], VerticeNormals[6], VerticeNormals[4]);
+            Polygons[3].SetLights(VerticeNormals[1], VerticeNormals[3], VerticeNormals[7], VerticeNormals[5]);
+            Polygons[4].SetLights(VerticeNormals[2], VerticeNormals[3], VerticeNormals[7], VerticeNormals[6]);
+            Polygons[5].SetLights(VerticeNormals[4], VerticeNormals[5], VerticeNormals[7], VerticeNormals[6]);
 
             Edges = new Edge[] 
             {
@@ -64,8 +94,15 @@ namespace VirtualCamera.Src
             };
         }
 
+        private Vector3 MeanNormal(Vector3 n1, Vector3 n2, Vector3 n3)
+        {
+            return new Vector3((n1.X + n2.X + n3.X) / 3, (n1.Y + n2.Y + n3.Y) / 3, 1);
+        }
+
         public override void Draw()
         {
+            //GraphicsManager.spriteBatch.Draw(crazy_cat, new Rectangle(100, 100, 256, 256), Color.White);
+
             if (IsOut())
             {
                 return;
